- Tweaked the lod help output a little. - Got rid of some MSVC compile warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2010/lod@34829 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lod/LodGen.h
@@ -104,13 +104,13 @@ public:
   size_t GetTriangleCount() const { return ordered_tris.GetSize(); }
 
   /// Get a triangle from the processed mesh
-  const csTriangle& GetTriangle(int i) const { return ordered_tris[i]; }
+  const csTriangle& GetTriangle(size_t i) const { return ordered_tris[i]; }
 
   /// Get the number of sliding windows in the processed mesh
   size_t GetSlidingWindowCount() const { return k.sliding_windows.GetSize(); }
 
   /// Get a sliding window from the processed mesh
-  const SlidingWindow& GetSlidingWindow(int i) const { return k.sliding_windows[i]; }
+  const SlidingWindow& GetSlidingWindow(size_t i) const { return k.sliding_windows[i]; }
   
 protected:
   /**
